Refactor pyenv domain knowledge out of the locator
