+++ src/client/pythonEnvironments/discovery/locators/services/posixKnownPathsLocator.ts
@@ -5,7 +5,7 @@ import { traceError } from '../../../../common/logger';
 import { PythonEnvKind, PythonEnvSource } from '../../../base/info';
 import { BasicEnvInfo, IPythonEnvsIterator, Locator } from '../../../base/locator';
 import { commonPosixBinPaths, getPythonBinFromPosixPaths } from '../../../common/posixUtils';
-import { isPyenvShimDir } from './pyenvLocator';
+import { isPyenvShimDir } from './pyenv';
 
 export class PosixKnownPathsLocator extends Locator<BasicEnvInfo> {
     private kind: PythonEnvKind = PythonEnvKind.OtherGlobal;
